+++ src/Monolog/Handler/DesktopNotificationHandler.php
@@ -1,0 +1,50 @@
+<?php
+
+namespace OctoLab\Common\Monolog\Handler;
+
+use Joli\JoliNotif;
+use Monolog\Handler\AbstractProcessingHandler;
+use Monolog\Logger;
+
+/**
+ * @author Kamil Samigullin <kamil@samigullin.info>
+ */
+class DesktopNotificationHandler extends AbstractProcessingHandler
+{
+    /** @var string */
+    private $name;
+    /** @var JoliNotif\Notifier|null */
+    private $notifier;
+    /** @var JoliNotif\Notification */
+    private $notification;
+
+    /**
+     * @param string $name
+     * @param int $level
+     * @param bool $bubble
+     */
+    public function __construct($name, $level = Logger::DEBUG, $bubble = true)
+    {
+        parent::__construct($level, $bubble);
+        $this->name = $name;
+        $this->notifier = JoliNotif\NotifierFactory::create();
+        $this->notification = new JoliNotif\Notification();
+    }
+
+    /**
+     * @param array $record
+     *
+     * @return bool
+     *
+     * @throws \Joli\JoliNotif\Exception\InvalidNotificationException
+     */
+    protected function write(array $record)
+    {
+        return $this->notifier && $this->notifier->send(
+            $this->notification
+                ->setTitle(sprintf('[%s] %s', Logger::getLevelName($this->level), $this->name))
+                // https://github.com/jolicode/JoliNotif/issues/24
+                ->setBody(str_replace('"', '', $record['formatted']))
+        );
+    }
+}